requirements.txt
# Python 3.12
python-dotenv>=1.1,<2
langchain[google-genai]==0.3.27